+++ src/LibHac/FsSystem/IntegrityVerificationStorage.cs
@@ -3,7 +3,7 @@ using System.Buffers;
 using System.IO;
 using LibHac.Crypto;
 using LibHac.Fs;
-using LibHac.FsSystem.Save;
+using LibHac.Tools.FsSystem.Save;
 
 namespace LibHac.FsSystem;
