+++ src/LibHac/Tools/FsSystem/Save/DuplexStorage.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using LibHac.Fs;
+using LibHac.FsSystem;
 
-namespace LibHac.FsSystem.Save;
+namespace LibHac.Tools.FsSystem.Save;
 
 public class DuplexStorage : IStorage
 {
