+++ src/LibHac/Tools/FsSystem/Save/JournalMap.cs
@@ -1,8 +1,9 @@
 ﻿using System.IO;
 using LibHac.Fs;
+using LibHac.FsSystem;
 using LibHac.Util;
 
-namespace LibHac.FsSystem.Save;
+namespace LibHac.Tools.FsSystem.Save;
 
 public class JournalMap
 {
