+++ src/LibHac/Tools/FsSystem/Save/SaveDataFile.cs
@@ -4,7 +4,7 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSystem.Save;
+namespace LibHac.Tools.FsSystem.Save;
 
 public class SaveDataFile : IFile
 {
