+++ src/LibHac/Tools/FsSystem/Save/SaveDataFileSystem.cs
@@ -5,9 +5,10 @@ using LibHac.Common.Keys;
 using LibHac.Crypto;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
+using LibHac.FsSystem;
 using Path = LibHac.Fs.Path;
 
-namespace LibHac.FsSystem.Save;
+namespace LibHac.Tools.FsSystem.Save;
 
 public class SaveDataFileSystem : IFileSystem
 {
