+++ src/LibHac/Tools/FsSystem/Save/SaveDataFileSystemCore.cs
@@ -2,10 +2,11 @@
 using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
+using LibHac.FsSystem;
 using LibHac.Util;
 using Path = LibHac.Fs.Path;
 
-namespace LibHac.FsSystem.Save;
+namespace LibHac.Tools.FsSystem.Save;
 
 public class SaveDataFileSystemCore : IFileSystem
 {
