+++ src/LibHac/Tools/FsSystem/Save/SaveFsEntry.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 
-namespace LibHac.FsSystem.Save;
+namespace LibHac.Tools.FsSystem.Save;
 
 internal ref struct SaveEntryKey
 {
